rename datastorm.js to index.js in lib
